Refactored test simulations.
