+++ glotaran/builtin/io/folder/test/test_folder_plugin.py
@@ -8,7 +8,7 @@ import pytest
 from glotaran.analysis.optimize import optimize
 from glotaran.io import save_result
 from glotaran.project.result import Result
-from glotaran.testing.sequential_spectral_decay import SCHEME
+from glotaran.testing.simulated_data.sequential_spectral_decay import SCHEME
 
 
 @pytest.fixture(scope="session")
