+++ glotaran/testing/test/test_example.py
@@ -1,7 +1,7 @@
 import xarray as xr
 
-from glotaran.testing.parallel_spectral_decay import DATASET as parallel_dataset
-from glotaran.testing.sequential_spectral_decay import DATASET as sequential_dataset
+from glotaran.testing.simulated_data.parallel_spectral_decay import DATASET as parallel_dataset
+from glotaran.testing.simulated_data.sequential_spectral_decay import DATASET as sequential_dataset
 
 
 def test_dataset():
